再三彻查修复：2 Critical + 1 Major + Minor（第二轮深度排查）
== Critical 修复（双宿主下功能必然失败）==

C1. CoreRuntimeSite 缺 ExtraHeadersJson → Codex 双宿主转发 100% 失败
   - Codex 隐藏 Site 的自定义请求头（Originator/Chatgpt-Account-Id/User-Agent）
     存在 Site.ExtraHeadersJson 里，但没进 Core 配置快照
   - Core 转发时 MergeExtraHeaders 拿到空，Codex 请求被上游拒绝
   - 修复：CoreRuntimeSite 加 ExtraHeadersJson 字段；SnapshotBuilder/BuildPatchAsync
     投影该字段；Core 端 GetRouteTargetsAsync 填充 ExtraHeaders

C2. /health 被 FallbackPolicy 拦截 → 健康检查 401
   - 上次加的 FallbackPolicy 覆盖所有未标 [AllowAnonymous] 的端点，
     /health 没标 → 负载均衡/k8s 探针误判不健康
   - 修复：/health 加 .AllowAnonymous()

== Major 修复（功能静默失效）==

M1. UpdateConcurrency 不推 SiteModelMappings → Core 并发限制不生效
   - MaxConcurrency 存在 SiteModelMappings，但 UpdateConcurrency 只推
     Sites/RouteRules/RouteEntries（InvalidateRouteTargetsAsync），不含 SiteModelMappings
   - Core 端 ModelConcurrencyLimiter 读不到新值，前端改并发不生效
   - 修复：UpdateConcurrency 补调 InvalidateModelMetadataAsync（推 SiteModelMappings）

== Minor 清理 ==
- CoreConfigSyncController 删除重复的 using AITool.Infrastructure.Proxy
- AdminUsageLogSummaryDto 的 TotalTokens/MaxDurationMs 从 int 改为 long
  （闭环 commit 9391615 的 Int32 溢出修复，SQL 端已 long 但 DTO 端仍 int）

== 已知性能问题（本次不重构，记录备查）==
- CodexInspectionService/CodexCooldownRecoveryService 在 SerialExecuteAsync 锁内
  调 InvalidateRouteTargetsAsync（含 HTTP），Core 慢响应时阻塞后台 DB 写。
  CodexTokenRefreshService 已是正确模式（锁外失效），巡检服务待后续对齐。

测试：365/365 全部通过（Application 190 + Admin 11 + Core 164）

+++ src/AITool.Admin/Controllers/Admin/UsageLogsApiController.cs
@@ -396,14 +396,14 @@ public sealed class AdminUsageLogSummaryDto
     public double SuccessRate { get; set; }
 
     /// <summary>
-    /// Token 总数。
+    /// Token 总数（用 long 避免大窗口累计超 Int32.MaxValue 溢出）。
     /// </summary>
-    public int TotalTokens { get; set; }
+    public long TotalTokens { get; set; }
 
     /// <summary>
     /// 最大耗时。
     /// </summary>
-    public int MaxDurationMs { get; set; }
+    public long MaxDurationMs { get; set; }
 }
 
 /// <summary>
@@ -526,8 +526,8 @@ public sealed class UsageLogsApiController : ControllerBase
             ? 0d
             : Math.Round(successRequests * 100d / totalRequests, 2, MidpointRounding.AwayFromZero);
 
-        var totalTokens = (int)(row.TotalTokens ?? 0);
-        var maxDurationMs = (int)(row.MaxDurationMs ?? 0);
+        var totalTokens = row.TotalTokens ?? 0;
+        var maxDurationMs = row.MaxDurationMs ?? 0;
 
         return Ok(new AdminUsageLogSummaryDto
         {
